Add files via upload
Once upon a time I was handed a mySQL database and asked to create a series of customer facing data entry screens. Boring. I wrote this as a started for 10. It's a PHP script which, given a table name, interrogates the metadata and generates the bare bones HTML form for the CRUD actions.